Move collections to separate file
