+++ Client/Frontend/Home/Wallpapers/WallpaperDataManager.swift
@@ -138,27 +138,9 @@ struct WallpaperDataManager {
     private func allSpecialCollections() -> [WallpaperCollection]? {
         var specialCollections = [WallpaperCollection]()
 
-        let houseExpiryDate = Calendar.current.date(
-            from: DateComponents(year: 2022, month: 5, day: 1))
-        let projectHouse = WallpaperCollection(
-            wallpaperFileNames: [WallpaperID(name: "trRed",
-                                             accessibilityLabel: "Turning Red wallpaper, giant red panda"),
-                                 WallpaperID(name: "trGroup",
-                                             accessibilityLabel: "Turning Red wallpaper, Mei and friends")],
-            ofType: .themed(type: .projectHouse),
-            expiringOn: houseExpiryDate,
-            limitedToLocales: ["en_US", "es_US"])
 
-        specialCollections.append(projectHouse)
-
-        let v100Celebration = WallpaperCollection(
-            wallpaperFileNames: [WallpaperID(name: "beachVibes",
-                                             accessibilityLabel: accessibilityIDs.FxBeachHillsWallpaper),
-                                 WallpaperID(name: "twilightHills",
-                                             accessibilityLabel: accessibilityIDs.FxTwilightHillsWallpaper)],
-            ofType: .themed(type: .v100Celebration))
-
-        specialCollections.append(v100Celebration)
+        specialCollections.append(projectHouseCollection())
+        specialCollections.append(v100CelebrationCollection())
 
         return specialCollections.isEmpty ? nil : specialCollections
     }
